requirements.txt
flask
flask-socketio
opencv-python
numpy
deepface
tf-keras
gTTS           
google-genai    